requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Google Gemini AI
google-genai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Streamlit (for the original app.py if you want to use it)
streamlit>=1.30.0
nltk>=3.8.0
tqdm>=4.65.0
scikit-learn>=1.3.0
